Fix for the whitelabel error
FIxed whitelabel error.
Added a package for controllers.
